backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "SlideFlip Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "websockets",
    "python-multipart",
    "aiofiles",
    "python-dotenv",
    "pydantic",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-socketio",
    "aiohttp",
    "asyncio-mqtt",
    "beautifulsoup4",
    "openai",
    "python-pptx",
    "Pillow",
    "networkx",
    "scikit-learn",
    "numpy",
    "transformers",
    "pydantic-settings>=2.10.1",
    "tiktoken>=0.11.0",
    "tavily-python>=0.7.10",
    "firecrawl-py>=2.16.5",
    "pdfminer.six",
    "python-docx",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.2.0",
    "pypdf>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
